StarterKit/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
minio==7.2.5
pandas==2.2.1
